refactor: migrate no-priors sampler
