+++ examples/no-priors-characterization/example_yamls/op_quick_exploration.yaml
@@ -1,13 +1,26 @@
 # Copyright IBM Corporation 2025, 2026
 # SPDX-License-Identifier: MIT
+# Baseline random ordering using random_walk with no-priors sampler
 spaces:
   - space-c8717f-3a68bf
 operation:
   module:
-    operationType: characterize
-    operatorName: no_priors_characterization
+    operationType: explore
+    operatorName: random_walk
   parameters:
-    targetOutput: yield
-    samples: 20
+    numberEntities: 20
     batchSize: 5
-    sampling_strategy: random
+    samplerConfig:
+      module:
+        moduleName: orchestrator.core.discoveryspace.no_priors_sampler
+        moduleClass: NoPriorsSampleSelector
+      parameters:
+        targetOutput: yield
+        samples: 20
+        batchSize: 5
+        sampling_strategy: random
+    singleMeasurement: true
+    filter:
+      filterMode: unmeasured
+
+# Made with Bob
