--- plugins/operators/no-priors-characterization/pyproject.toml
@@ -1,29 +1,0 @@
-[project]
-name = "ado-no-priors-characterization"
-description = "No-priors characterization operator for sampling discovery spaces using high-dimensional sampling strategies"
-readme = "README.md"
-requires-python = ">=3.10,<3.14"
-dependencies = [
-  "ado-core",
-  "numpy",
-  "pandas>=2.2.0",
-  "scipy",
-]
-dynamic = ["version"]
-
-[project.entry-points."ado.operators"]
-no-priors-characterization = "no_priors_characterization.operator"
-
-[build-system]
-requires = ["setuptools", "setuptools_scm"]
-build-backend = "setuptools.build_meta"
-
-[tool.setuptools.packages.find]
-where = ["src"]
-include = ["no_priors_characterization*"]
-
-[tool.setuptools_scm]
-root = "../../../"
-
-[tool.uv.sources]
-ado-core = { workspace = true }
