--- plugins/operators/no-priors-characterization/src/no_priors_characterization/__init__.py
@@ -1,6 +1,0 @@
-# Copyright IBM Corporation 2025, 2026
-# SPDX-License-Identifier: MIT
-
-from no_priors_characterization.operator import no_priors_characterization
-
-__all__ = ["no_priors_characterization"]
